- Added fixes for compiler warnings on a Mac

+++ library/debug.c
@@ -139,8 +139,9 @@ void debug_print_mpi( const ssl_context *ssl, int level,
         if( X->p[n] != 0 )
             break;
 
-    snprintf( str, maxlen, "%s(%04d): value of '%s' (%d bits) is:\n",
-              file, line, text, ((n + 1) * sizeof( t_int )) << 3 );
+    snprintf( str, maxlen, "%s(%04d): value of '%s' (%lu bits) is:\n",
+              file, line, text, 
+              (unsigned long) ((n + 1) * sizeof( t_int )) << 3 );
 
     str[maxlen] = '\0';
     ssl->f_dbg( ssl->p_dbg, level, str );
